Revise project description tagline in README.md
JavaScript permits a variety of different approaches to Object-Oriented programming and I have chosen to experiment with what Crockford refers to as a 'pseudoclassical' OO style. This does not reflect a preference. Rather it is an experiment in writing library code in a style that is endorsed by enough active programmers to justify knowing it well.